Create BrinkAJ - CSS update
I just made a minor update to your css that changes the nav word color on hover. I think its good to have that.  You can change it to what color you want in the .navbar li ul li a:hover:  but other than that I think it looks good.